pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bpy>=4.5.4",
    "fpsample>=0.3.3",
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "trimesh[all]>=4.9.0",
    "tqdm>=4.66.0",
    "scipy>=1.16.3",
    "scikit-image>=0.25.2",
    "shapely>=2.1.2",
]
